requirements.txt
streamlit~=1.13.0
Pillow==9.2.0